requirements.txt
PyYAML===6.0.1
selenium===4.11.2
selenium-wire===5.1.0
brotli===1.0.9
bs4===0.0.1
